Refactor organization of leaderboard classes
