+++ app/src/main/java/com/qrpokemon/qrpokemon/LeaderboardActivity.java
@@ -1,40 +1,31 @@
 package com.qrpokemon.qrpokemon;
-import android.content.Context;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
-import android.view.ViewGroup;
 import android.widget.AdapterView;
-import android.widget.BaseAdapter;
 import android.widget.ImageButton;
 import android.widget.ListView;
 import android.widget.Spinner;
-import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Observable;
-import java.util.Observer;
-
 public class LeaderboardActivity extends AppCompatActivity {
     // Init variables
-    LeaderboardController leaderboardController;
     private ListView leaderboardListView;
     private Spinner sortBy;
     private ImageButton backButton;
+    private int sortMethod = 0;
+
+    // Leaderboard variables
+    LeaderboardController leaderboardController;
     private LeaderboardList leaderboardList;
     private LeaderboardAdapter leaderboardAdapter;
-    private int sortMethod = 0;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.leaderboard_activity);
 
-        // TODO: Move controller functions into controller class
         // data binding
         leaderboardListView = findViewById(R.id.leaderboard_listview);
         sortBy = findViewById(R.id.sp_sort_selection);
@@ -44,7 +35,6 @@ public class LeaderboardActivity extends AppCompatActivity {
         leaderboardList = new LeaderboardList();
         leaderboardController = LeaderboardController.getInstance();
         leaderboardController.getLeaderboard(this, leaderboardList);
-        Log.d("Leaderboard: ", leaderboardList.getList().toString());
 
         // set adapter
         leaderboardAdapter = new LeaderboardAdapter(this, leaderboardList.getList());
@@ -55,177 +45,22 @@ public class LeaderboardActivity extends AppCompatActivity {
         sortBy.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
             @Override
             public void onItemSelected(AdapterView<?> adapterView, View view, int i, long l) {
-                // TODO: Use dropdown text String in switch statement?
                 // 0 = i = sortMethod( one Qrcode score)
                 // 1 = i = sortMethod( total Qrcode score)
                 // 2 = i = sortMethod( numbers of  Qrcode score)
+                sortMethod = i;  // FIXME: This assignment is local scoped
+            }
 
-                sortMethod = i;
-            }
             @Override
-            public void onNothingSelected(AdapterView<?> adapterView) {
-            }
+            public void onNothingSelected(AdapterView<?> adapterView) {}
         });
 
         // set click event for back button
         backButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                finish();
+                finish();  // Return to calling activity
             }
         });
     }
-}
-
-/**
- * Create an object for each item in leaderboard
- */
-class LeaderboardItem {
-    private int rank;
-    private String username;
-    private int highestScore;
-    private long qrCount;
-    private long totalScore;
-
-    public LeaderboardItem(String username, long qrCount, long totalScore) {
-        this.username = username;
-        this.highestScore = highestScore;
-        this.qrCount = qrCount;
-        this.totalScore = totalScore;
-    }
-
-    // TODO: Remove unused getters and setters
-    public int getRank() {
-        return rank;
-    }
-
-    public void setRank(int rank) {
-        this.rank = rank;
-    }
-
-    public String getUserName() {
-        return username;
-    }
-
-    public void setUserName(String username) {
-        this.username = username;
-    }
-
-    public int getHighestScore() {
-        return highestScore;
-    }
-
-    public void setHighestScore(int highestScore) {
-        this.highestScore = highestScore;
-    }
-
-    // TODO: Remove since we should just do another database fetch in this case
-    public long getQrQuantity() {
-        return qrCount;
-    }
-
-    public void setQrQuantity(int qrQuantity) {
-        this.qrCount = qrQuantity;
-    }
-
-    public long getTotalScore() {
-        return totalScore;
-    }
-
-    public void setTotalScore(int totalScore) {
-        this.totalScore = totalScore;
-    }
-}
-
-class LeaderboardList extends Observable {
-    private ArrayList<LeaderboardItem> list;
-
-    LeaderboardList() {
-        list = new ArrayList<>();
-    }
-
-    public void add(LeaderboardItem leaderboardItem) {
-        list.add(leaderboardItem);
-    }
-
-    public void notifyListUpdate() {
-        setChanged();
-        notifyObservers();
-    }
-
-    public List<LeaderboardItem> getList() {
-        return list;
-    }
-
-    public void sort(int sortMethod) {
-        switch (sortMethod) {
-            case 0:
-                list.sort(new LeaderboardCompareTotalScore());
-        }
-
-        notifyListUpdate();
-    }
-}
-
-// set adapter for leaderboard listview
-class LeaderboardAdapter extends BaseAdapter implements Observer {
-    private Context context;
-    private List<LeaderboardItem> leaderboardList;
-
-    public LeaderboardAdapter(Context context, List<LeaderboardItem> leaderboardList) {
-        this.context = context;
-        this.leaderboardList = leaderboardList;
-    }
-
-    @Override
-    public int getCount() {
-        return leaderboardList.size();
-    }
-
-    @Override
-    public Object getItem(int i) {
-        return null;
-    }
-
-    @Override
-    public long getItemId(int i) {
-        return 0;
-    }
-
-    @Override
-    public View getView(int i, View view, ViewGroup viewGroup) {
-        View view1 = View.inflate(context, R.layout.leaderboard_list,null);
-
-        TextView rankOrderList = view1.findViewById(R.id.rand_order_list);
-        TextView userNameList = view1.findViewById(R.id.username_list);
-        TextView highestScoreList = view1.findViewById(R.id.highestscore_list);
-        TextView qrNumList = view1.findViewById(R.id.qr_quantity_list);
-        TextView totalScoreList = view1.findViewById(R.id.total_score_list);
-
-        String a = String.valueOf(22);
-
-        rankOrderList.setText(String.valueOf(leaderboardList.get(i).getRank()));
-        userNameList.setText(String.valueOf(leaderboardList.get(i).getUserName()));
-        highestScoreList.setText(String.valueOf(leaderboardList.get(i).getHighestScore()));
-        qrNumList.setText(String.valueOf(leaderboardList.get(i).getQrQuantity()));
-        totalScoreList.setText(String.valueOf(leaderboardList.get(i).getTotalScore()));
-
-        return view1;
-    }
-
-    @Override
-    public void update(Observable observable, Object o) {
-        if (observable instanceof LeaderboardList)
-            // Update the list
-            leaderboardList = ((LeaderboardList) observable).getList();
-
-        this.notifyDataSetChanged();
-    }
-}
-
-// NOTE: Replace if Database can do the sorting
-class LeaderboardCompareTotalScore implements Comparator<LeaderboardItem> {
-        public int compare(LeaderboardItem first, LeaderboardItem second) {
-            return (int) (second.getTotalScore() - first.getTotalScore());
-        }
 }
